SDK-1661: Rename: dynamic_sharing_service -> dynamic, share -> extra
Multiple packages respectively fall under the "share" namespace, in one way or another, so we decided to remove this and have separate packages on the root
