+++ extra/extra_data.go
@@ -1,4 +1,4 @@
-package share
+package extra
 
 import (
 	"github.com/getyoti/yoti-go-sdk/v3/profile/attribute"
