+++ sh/go-build-modtidy.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env bash
 go build ./...
 
-for d in _examples/*/
-do
-     (cd "$d" && go mod tidy)
+for d in _examples/*/; do
+  (cd "$d" && go mod tidy)
 done
